moved auth to BasicMethods
