+++ tests/BasicMethods.py
@@ -1,0 +1,75 @@
+# -*- coding: utf-8 -*-
+import os
+
+import unittest
+import urlparse
+
+from selenium.webdriver import DesiredCapabilities, Remote
+from selenium.webdriver.support.ui import WebDriverWait
+
+
+class Page(object):
+    BASE_URL = 'https://ok.ru/'
+    PATH = ''
+
+    def __init__(self, driver):
+        self.driver = driver
+
+    def open(self):
+        url = urlparse.urljoin(self.BASE_URL, self.PATH)
+        self.driver.get(url)
+        self.driver.maximize_window()
+
+
+class Component(object):
+    def __init__(self, driver):
+        self.driver = driver
+
+
+class AuthPage(Page):
+    PATH = ''
+
+    @property
+    def form(self):
+        return AuthForm(self.driver)
+
+
+class AuthForm(Component):
+    LOGIN = '//input[@name="st.email"]'
+    PASSWORD = '//input[@name="st.password"]'
+    SUBMIT = '//input[@class="button-pro form-actions_yes"]'
+
+    def set_login(self, login):
+        self.driver.find_element_by_xpath(self.LOGIN).send_keys(login)
+
+    def set_password(self, pwd):
+        self.driver.find_element_by_xpath(self.PASSWORD).send_keys(pwd)
+
+    def submit(self):
+        self.driver.find_element_by_xpath(self.SUBMIT).click()
+
+    
+
+class GroupMessagesTest(unittest.TestCase):
+    def setUp(self):
+        browser = 'FIREFOX'
+
+        self.driver = Remote(
+            command_executor='http://127.0.0.1:4444/wd/hub',
+            desired_capabilities=getattr(DesiredCapabilities, browser).copy()
+        )
+
+    def tearDown(self):
+        self.driver.quit()
+
+    def test(self):
+        auth_page = AuthPage(self.driver)
+        auth_page.open()
+
+        auth_form = auth_page.form
+        auth_form.set_login('technopark16')
+        auth_form.set_password('testQA1')
+        auth_form.submit()
+
+        user_name = auth_page.top_menu.get_username()
+        self.assertEqual('technopark16 technopark16', user_name)
