+++ tests/group_messages_test.py
@@ -3,63 +3,10 @@ import os
 
 import unittest
 import urlparse
+from BasicMethods import Page, Component, AuthPage
 
 from selenium.webdriver import DesiredCapabilities, Remote
 from selenium.webdriver.support.ui import WebDriverWait
-
-
-class Page(object):
-    BASE_URL = 'https://ok.ru/'
-    PATH = ''
-
-    def __init__(self, driver):
-        self.driver = driver
-
-    def open(self):
-        url = urlparse.urljoin(self.BASE_URL, self.PATH)
-        self.driver.get(url)
-        self.driver.maximize_window()
-
-
-class Component(object):
-    def __init__(self, driver):
-        self.driver = driver
-
-
-class AuthPage(Page):
-    PATH = ''
-
-    @property
-    def form(self):
-        return AuthForm(self.driver)
-
-    @property
-    def top_menu(self):
-        return TopMenu(self.driver)
-
-
-class TopMenu(Component):
-    USERNAME = '//h1[@class="mctc_name_tx"]'
-
-    def get_username(self):
-        return WebDriverWait(self.driver, 30, 0.1).until(
-            lambda d: d.find_element_by_xpath(self.USERNAME).text
-        )
-
-
-class AuthForm(Component):
-    LOGIN = '//input[@name="st.email"]'
-    PASSWORD = '//input[@name="st.password"]'
-    SUBMIT = '//input[@class="button-pro form-actions_yes"]'
-
-    def set_login(self, login):
-        self.driver.find_element_by_xpath(self.LOGIN).send_keys(login)
-
-    def set_password(self, pwd):
-        self.driver.find_element_by_xpath(self.PASSWORD).send_keys(pwd)
-
-    def submit(self):
-        self.driver.find_element_by_xpath(self.SUBMIT).click()
 
 
 class GroupMessagesTest(unittest.TestCase):
@@ -83,5 +30,4 @@ class GroupMessagesTest(unittest.TestCase):
         auth_form.set_password('testQA1')
         auth_form.submit()
 
-        user_name = auth_page.top_menu.get_username()
-        self.assertEqual('technopark16 technopark16', user_name)
+        self.assertEqual('technopark16 technopark16', '123')
